requirements.txt
black==22.3.0
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pymongo==4.1.1
PyYAML==6.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.1
zipp==3.8.0
